refactor(backend/common): put logging setup and exceptions in common files
share with mqtt and api
